[assembly-preparer] Refactor the saving of assemblies to a separate step.
