+++ tools/dotnet-linker/LinkerConfiguration.cs
@@ -10,6 +10,7 @@ using Mono.Cecil;
 using Mono.Linker;
 using Mono.Linker.Steps;
 
+using Xamarin.Build;
 using Xamarin.Bundler;
 using Xamarin.Utils;
 using Xamarin.Tuner;
@@ -88,6 +89,7 @@ namespace Xamarin.Linker {
 
 #if ASSEMBLY_PREPARER
 		public List<AssemblyDefinition> Assemblies => Application.LinkContext.Assemblies;
+		public required List<AssemblyPreparerInfo> AssemblyInfos;
 		public List<(string Path, AssemblyDefinition Assembly, string? OriginatingAssembly)> AddedAssemblies = new ();
 #else
 		// The list of assemblies is populated in CollectAssembliesStep.
